Refactoed and split into modules
