+++ mood-tracker.cabal
@@ -27,6 +27,10 @@ common shared
     -Wmissing-deriving-strategies -Wunused-foralls -Wunused-foralls
     -fprint-explicit-foralls -fprint-explicit-kinds
 
+  other-modules:
+    MoodTracker.HTML
+    MoodTracker.Parser
+
   mixins:
     base hiding (Prelude),
     relude (Relude as Prelude, Relude.Container.One),
@@ -103,3 +107,13 @@ common shared
 executable mood-tracker
   import:  shared
   main-is: Main.hs
+
+--test-suite my-tests
+--  import:         shared
+--  main-is:        Spec.hs
+--  type:           exitcode-stdio-1.0
+--  hs-source-dirs: tests
+--  other-modules:  MainSpec
+--  build-depends:
+--    , hspec
+--    , hspec-discover
